fix: Corrige problemas na estrutura do projeto e remoção de arquivos
